Moved Image into the base module
