+++ examples/em-test.cpp
@@ -4,7 +4,7 @@
 #include <stdlib.h>
 #include <math.h>
 #include "em/base/error.h"
-#include "em/image/image.h"
+#include "em/base/image.h"
 #include "em/base/type.h"
 #include  <map>
