+++ include/em/base/image.h
