+++ src/base/image_formats/image_em.cpp
@@ -1,7 +1,7 @@
 
 #include "em/base/error.h"
-#include "em/image/image.h"
-#include "em/image/image_priv.h"
+#include "em/base/image.h"
+#include "em/base/image_priv.h"
 
 using namespace em;
