+++ src/base/image_formats/image_mrc.cpp
@@ -1,8 +1,8 @@
 #include <iomanip>
 
 #include "em/base/error.h"
-#include "em/image/image.h"
-#include "em/image/image_priv.h"
+#include "em/base/image.h"
+#include "em/base/image_priv.h"
 
 using namespace em;
