+++ src/base/image_formats/image_tiff.cpp
@@ -5,8 +5,8 @@
 #include "tiffio.h"
 
 #include "em/base/error.h"
-#include "em/image/image.h"
-#include "em/image/image_priv.h"
+#include "em/base/image.h"
+#include "em/base/image_priv.h"
 
 using namespace em;
